Backport all bug fixes and improvements
Also fix Washcloth dupe bug

+++ src/main/java/com/shinoow/abyssalcraft/api/recipe/EngraverRecipes.java
@@ -5,7 +5,7 @@
  * are made available under the terms of the GNU Lesser Public License v3
  * which accompanies this distribution, and is available at
  * http://www.gnu.org/licenses/lgpl-3.0.txt
- * 
+ *
  * Contributors:
  *     Shinoow -  implementation
  ******************************************************************************/
@@ -110,7 +110,7 @@ public class EngraverRecipes {
 
 		return experienceList.entrySet().stream()
 				.filter(e -> areStacksEqual(stack, e.getKey()))
-				.map(e -> e.getValue().floatValue())
+				.map(e -> e.getValue())
 				.findFirst()
 				.orElse(0.0F);
 	}
