+++ src/main/java/com/shinoow/abyssalcraft/common/blocks/itemblock/ItemShoggothOoze.java
@@ -5,7 +5,7 @@
  * are made available under the terms of the GNU Lesser Public License v3
  * which accompanies this distribution, and is available at
  * http://www.gnu.org/licenses/lgpl-3.0.txt
- * 
+ *
  * Contributors:
  *     Shinoow -  implementation
  ******************************************************************************/
@@ -56,11 +56,11 @@ public class ItemShoggothOoze extends ItemBlockAC
 
 			if (block == this.block)
 			{
-				int i = iblockstate.getValue(BlockShoggothOoze.LAYERS).intValue();
+				int i = iblockstate.getValue(BlockShoggothOoze.LAYERS);
 
 				if (i <= 7)
 				{
-					IBlockState iblockstate1 = iblockstate.withProperty(BlockShoggothOoze.LAYERS, Integer.valueOf(i + 1));
+					IBlockState iblockstate1 = iblockstate.withProperty(BlockShoggothOoze.LAYERS, i + 1);
 					AxisAlignedBB axisalignedbb = iblockstate1.getCollisionBoundingBox(worldIn, blockpos);
 
 					if (axisalignedbb != Block.NULL_AABB && worldIn.checkNoEntityCollision(axisalignedbb.offset(blockpos)) && worldIn.setBlockState(blockpos, iblockstate1, 10))
